Fix review commentaries
