+++ src/components/DynamicTable/TableAnimation.ts
