+++ src/components/DynamicTable/TableAnimations.css
