+++ src/components/ResultsTable/ResultsTable.tsx
@@ -1,12 +1,12 @@
 import React, { useRef, useEffect, useState } from "react";
-import { DynamicTable } from "../DinamicTable/DynamicTable";
+import { DynamicTable } from "../DynamicTable/DynamicTable";
 import { ColumnDefinition } from "../../Types/Componets/DynamicTableTypes";
 import { IProperty } from "../../Types/Models/Property";
 import { IResultsTableProps } from "../../Types/Componets/IResultsComponentsProps";
 import { AdjustmentTypeEnum } from "../../Models/Enums/AdjustmentTypeEnum";
 import { IAdjustment } from "../../Types/Models/Adjustment";
 import "./ResultsTable.css";
-import LiquidationMethodEnum from "@/Models/Enums/LiquidationMethodEnum";
+import { LiquidationMethodEnum } from "../../Models/Enums/LiquidationMethodEnum";
 
 export function ResultsTable({ properties }: IResultsTableProps){
     // Column definitions for the results table
