+++ vite.config.js
@@ -4,7 +4,7 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [react()],
   esbuild: {
-    jsxInject: `import React from 'react'`,
+
     jsx: 'automatic'
   },
   resolve: {
